move sampling code to sampler.cc
